requirements.txt
alembic==0.9.8
Babel==2.5.3
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.2.1
defusedxml==0.5.0
dominate==2.3.1
elasticsearch==6.2.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
flipflop==1.0
guess-language==0.2
guess-language-spirit==0.5.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pbr==3.1.1
PyJWT==1.6.1
python-dateutil==2.7.0
python-dotenv==0.8.2
python-editor==1.0.3
python3-openid==3.1.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
rq==0.10.0
six==1.11.0
SQLAlchemy==1.2.5
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
Tempita==0.5.2
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.1
